dynamic allocation of a struct chunk
Allocation of the chunk structure and its addition to the map is
performed while holding the snapshot as needed. This optimization
allows to use small chunks even on large block devices.

+++ drivers/block/blksnap/diff_area.h
@@ -93,7 +93,7 @@ struct diff_area {
 	struct block_device *orig_bdev;
 	struct diff_storage *diff_storage;
 
-	unsigned long long chunk_shift;
+	unsigned long chunk_shift;
 	unsigned long chunk_count;
 	struct xarray chunk_map;
