+++ drivers/block/blksnap/params.h
@@ -8,7 +8,7 @@ int get_tracking_block_maximum_shift(void);
 int get_tracking_block_maximum_count(void);
 int get_chunk_minimum_shift(void);
 int get_chunk_maximum_shift(void);
-int get_chunk_maximum_count(void);
+unsigned long get_chunk_maximum_count(void);
 int get_chunk_maximum_in_queue(void);
 int get_free_diff_buffer_pool_size(void);
 int get_diff_storage_minimum(void);
